+++ Tests/EditMode/SimultriaUnityBuildVersionLookupTests.cs
@@ -4,6 +4,7 @@ using Deucarian.API.Models;
 using Deucarian.Simultria.API.Configuration;
 using Deucarian.Simultria.API.Models;
 using Deucarian.Simultria.API.Services;
+using Deucarian.Simultria.UnityBuildRouting;
 using Newtonsoft.Json;
 using NUnit.Framework;